langgraph_study/requirements.txt
langgraph==0.1.14
langchain==0.2.14
langchain-openai==0.1.19
python-dotenv==1.0.0
matplotlib==3.7.1
networkx==3.1